Fix some styling, add back and forward navigation in app, add navigation to dashboard. Weather locations can be deleted

+++ src/main/java/com/jorgma/homeapp/weather/dao/WeatherLocationRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 /**
  * Created by jorgma on 2018-02-21.
  */
-public interface WeatherLocationRepository extends PagingAndSortingRepository<WeatherLocation, Long> {
+public interface WeatherLocationRepository extends PagingAndSortingRepository<WeatherLocation, Integer> {
 
     Optional<WeatherLocation> findByCurrent(boolean isCurrent);
 }
